Add collation to Cldr.HTML.Territories.select/4 Currently the collation is done on the territory info map which means it uses erlang term ordering which is not what is expected by the user.
The commit adds a `:collator` option to sort the
territories with a default collator that sorts
by territory name.

Since the default collator sorts the names using
`<` comparison, it does not sort in a Unicode-aware
manner. Therefore default collation will not return
results that are expected by users in locales that
do not use the ASCII character set.